+++ src/main/java/com/bcam/bcmonitor/api/DashController.java
@@ -60,7 +60,7 @@ public class DashController {
     }
 
     @GetMapping("/blockhash/{height}")
-    Mono<String> getBestBlockHash(@PathVariable int height) {
+    Mono<String> getBestBlockHash(@PathVariable long height) {
         return client.getBlockHash(height);
     }
